refactor: Make bowlresult package
