+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,9 +1,9 @@
 package bowling.domain.frame;
 
-import static bowling.domain.BonusResult.NULL_RESULT;
+import static bowling.domain.bowlresult.BonusResult.NULL_RESULT;
 
-import bowling.domain.BonusResult;
-import bowling.domain.RegularResult;
+import bowling.domain.bowlresult.BonusResult;
+import bowling.domain.bowlresult.RegularResult;
 import bowling.domain.Round;
 import bowling.domain.Score;
 import bowling.domain.framestate.FrameState;
