+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.RegularResult;
+import bowling.domain.bowlresult.RegularResult;
 import bowling.domain.Round;
 import bowling.domain.Score;
 import bowling.exception.CannotBowlException;
